get testing to work
