+++ frontend/ui/lib/widgets/create_trip_widget.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:tripmeout/widgets/autocomplete_text_field_widget.dart';
 import 'package:tripmeout/services/trip_service.dart';
 import 'package:tripmeout/services/places_services.dart';
 import 'package:tripmeout/model/trip.dart';
