+++ frontend/ui/test/widgets/create_trip_widget_test.dart
@@ -6,11 +6,13 @@ import 'package:tripmeout/widgets/create_trip_widget.dart';
 import 'package:tripmeout/widgets/map_widget.dart';
 import 'package:tripmeout/services/trip_service.dart';
 import 'package:tripmeout/model/trip.dart';
+import 'package:tripmeout/widgets/autocomplete_text_field_widget.dart';
 
 class MockTripService extends Mock implements TripService {}
+class MockPlacesApiServices extends Mock implements PlacesApiServices {}
 
 void main() {
-  final PlacesApiServices placesApiServices = PlacesApiServices();
+  final PlacesApiServices placesApiServices = MockPlacesApiServices();
 
   testWidgets('Showing the text on page correctly shows up',
       (WidgetTester tester) async {
@@ -22,11 +24,12 @@ void main() {
     await tester.pumpAndSettle();
 
     // Should be everything on the create trip screen...
-    expect(find.text('Enter your Location'), findsOneWidget);
-    expect(find.text('Radius KM'), findsOneWidget);
+    expect(find.text('enter your trip name'), findsOneWidget);
+    expect(find.text('Enter your Destination'), findsOneWidget);
     // TODO: Enable this when/if a MapWidget is added?
     // expect(find.byType(MapWidget), findsOneWidget);
     expect(find.byType(RaisedButton), findsOneWidget);
+    expect(find.byType(MapsApiPlacesTextFieldWidget), findsOneWidget);
   });
 
   testWidgets('Testing the submit button which creates a trip',
@@ -40,22 +43,17 @@ void main() {
     await tester.pumpWidget(wrapForDirectionality(createTripsWidget));
     await tester.pumpAndSettle();
 
-    // TODO: Consider enabling this after fixing the test/code. The display
-    // behavior has changed since this was first introduced.
+    await tester.enterText(
+         find.widgetWithText(TextField, 'enter your trip name'), "Italy");
+    await tester.pumpAndSettle();
+    
+    await tester.tap(find.widgetWithText(RaisedButton, 'Submit'));
+    await tester.pumpAndSettle();
 
-    // await tester.enterText(
-    //     find.widgetWithText(TextField, 'Enter your Location'), "Italy");
-    // await tester.enterText(find.widgetWithText(TextField, 'Radius KM'), "25");
+    Trip createdTrip =
+         verify(tripService.createTrip(captureAny)).captured.single;
 
-    // RaisedButton button =
-    //     find.widgetWithText(RaisedButton, 'Submit').evaluate().first.widget;
-
-    // button.onPressed();
-    // await tester.pump();
-
-    // Trip createdTrip =
-    //     verify(tripService.createTrip(captureAny)).captured.single;
-    // expect(createdTrip.name, equals('Italy'));
+    expect(createdTrip.name, equals('Italy'));
   });
 }
